create readout affinity script that parses oks to infer the thread names and create the pinning
